refactor: GeneralSearchProperties를 config 패키지로 이동하고, 상수는 서비스 하나에서만 사용하므로 흡수
