+++ src/test/java/com/techfork/evaluation/search/setup/SearchGroundTruthGenerator.java
@@ -6,7 +6,7 @@ import com.techfork.domain.activity.repository.ScrabPostRepository;
 import com.techfork.domain.post.document.PostDocument;
 import com.techfork.domain.post.repository.PostRepository;
 import com.techfork.domain.search.dto.SearchResult;
-import com.techfork.domain.search.service.GeneralSearchProperties;
+import com.techfork.domain.search.config.GeneralSearchProperties;
 import com.techfork.domain.search.service.SearchServiceImpl;
 import com.techfork.domain.user.repository.UserProfileDocumentRepository;
 import com.techfork.evaluation.recommendation.setup.components.FileExporter;
